Flask/requirements.txt
flask>=3.0
tensorflow>=2.10
deepface>=0.0.93
opencv-python>=4.9
Pillow>=10.0
numpy>=1.24